+++ src/main/java/net/pointofviews/club/dto/response/ReadAllClubMembersResponse.java
@@ -10,18 +10,5 @@ public record ReadAllClubMembersResponse(
         @Schema(description = "클럽 멤버 목록")
         List<ClubMemberResponse> clubMember
 ) {
-    public record ClubMemberResponse(
-            @Schema(description = "이메일", example = "user@example.com")
-            String email,
 
-            @Schema(description = "닉네임", example = "nickname")
-            String nickname,
-
-            @Schema(description = "프로필 이미지", example = "https://example.com/image.jpg")
-            String profileImage,
-
-            @Schema(description = "리더 여부", example = "false")
-            boolean isLeader
-    ) {
-    }
 }
