+++ src/main/java/net/pointofviews/club/service/impl/MemberClubServiceImpl.java
@@ -2,6 +2,7 @@ package net.pointofviews.club.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import net.pointofviews.club.domain.MemberClub;
+import net.pointofviews.club.dto.response.ClubMemberResponse;
 import net.pointofviews.club.dto.response.ReadAllClubMembersResponse;
 import net.pointofviews.club.dto.response.ReadClubMemberListResponse;
 import net.pointofviews.club.dto.response.ReadClubMemberResponse;
@@ -70,7 +71,7 @@ public class MemberClubServiceImpl implements MemberClubService {
             throw ClubException.clubNotFound(clubId);
         }
 
-        List<ReadAllClubMembersResponse.ClubMemberResponse> allMembersByClubId = memberClubRepository.findAllMembersByClubId(clubId);
+        List<ClubMemberResponse> allMembersByClubId = memberClubRepository.findAllMembersByClubId(clubId);
         return new ReadAllClubMembersResponse(allMembersByClubId);
     }
 }
